requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
# PyMySQL==1.1.1 # Removido pois não é usado e vamos usar PostgreSQL
SQLAlchemy==2.0.40
cryptography==36.0.2
gunicorn==21.2.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
Werkzeug>=3.1
psycopg2-binary # Adicionado para conexão PostgreSQL
